feat(installer): streamline GTA 5 location process and globalize FiveMDirectory
This commit introduces a couple of significant changes to the installation script. Firstly, it removes the previous method of locating the GTA 5 directory since it is not required.

Additionally, the `FiveMDirectory` variable has been declared as GLOBAL to ensure its availability throughout the script, which improves the maintainability and readability of the code by using a consistent reference to the FiveM directory.

Lastly, a check for the current directory entry being "." is added in the reshade files copying loop to prevent unnecessary operations on the current directory reference.